February 12th
Been sick, done the highlight. HP system is now the last tricky part, kinda. Need add a camera animation to trigger when the health is 0, then restart scene. Oh and hitreg will be a pain in my ass.
After that it's smooth sailing with modeling chapter 2.
Thankfully a lot of this code is reusable so... :D